server/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asttokens==2.4.1
azure-cognitiveservices-speech==1.40.0
blinker==1.8.2
boto3==1.35.29
botocore==1.35.29
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
distro==1.9.0
einops==0.8.0
encodec==0.1.1
executing==2.1.0
filelock==3.16.1
Flask==3.0.3
fsspec==2024.9.0
funcy==2.0
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-texttospeech==2.17.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
openai==1.50.2
packaging==24.1
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
suno-bark @ git+https://github.com/suno-ai/bark.git@f4f32d4cd480dfec1c245d258174bc9bde3c2148
sympy==1.13.3
tokenizers==0.20.0
torch==2.4.1
torchaudio==2.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.1
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
